--- a/packages/turf-isobands/index.ts
+++ b/packages/turf-isobands/index.ts
@@ -19,9 +19,7 @@ import {
   Polygon,
   Feature,
 } from "geojson";
-
 import { gridToMatrix } from "./lib/grid-to-matrix.js";
-import { isoBands } from "marchingsquares";
 
 type GroupRingProps = { [prop: string]: string };
 type GroupedRings =
@@ -70,6 +68,11 @@ function isobands(
 
   // Isoband methods
   const matrix = gridToMatrix(pointGrid, { zProperty: zProperty, flip: true });
+
+  // A quick note on what 'top' and 'bottom' mean in coordinate system of `matrix`:
+  // Remember that the southern hemisphere is represented by negative numbers,
+  // so a matrix Y of 0 is actually the *bottom*, and a Y of dy - 1 is the *top*.
+
   let contours = createContourLines(matrix, breaks, zProperty);
   contours = rescaleContours(contours, matrix, pointGrid);
 
@@ -114,24 +117,399 @@ function createContourLines(
   property: string
 ): GroupedRings[] {
   const contours: GroupedRings[] = [];
-  for (let i = 1; i < breaks.length; i++) {
-    const lowerBand = +breaks[i - 1]; // make sure the breaks value is a number
-    const upperBand = +breaks[i];
 
-    const isobandsCoords = isoBands(matrix, lowerBand, upperBand - lowerBand);
+  let prevSegments: Position[][];
+  for (let i = 1; i < breaks.length; i++) {
+    // the first time through this loop, we need to create the segments for the first break
+    if (i === 1) {
+      prevSegments = getSegments(matrix, +breaks[0]);
+    }
+
+    const upperBand = +breaks[i]; // make sure the breaks value is a number
+    const lowerBand = +breaks[i - 1];
+    const segments = getSegments(matrix, upperBand);
+
+    // We will use breaks[i]'s rings to help close breaks[i-1]'s rings.
+    // breaks[i]'s rings are clockwise from the point of view of breaks[i - 1] and must be reversed for proper counterclockwise ordering.
+    // At the same time, we clone each Position, so that we don't use the same Position Array instance in different output geometries.
+    const reverseSegments = segments.map((segment) =>
+      // note that we (in-place) reverse the array result of .map and not the original segment itself.
+      segment.map((pos) => [pos[0], pos[1]]).reverse()
+    );
+
+    // use the segments from breaks[i-1] and breaks[i] to create rings, which will
+    // then be combined into polygons in the next steps.
+    const rings = assembleRings(prevSegments!.concat(reverseSegments), matrix);
+
     // as per GeoJson rules for creating a Polygon, make sure the first element
     // in the array of LinearRings represents the exterior ring (i.e. biggest area),
     // and any subsequent elements represent interior rings (i.e. smaller area);
     // this avoids rendering issues of the MultiPolygons on the map
-    const nestedRings = orderByArea(isobandsCoords);
-    const groupedRings = groupNestedRings(nestedRings);
+    const orderedRings = orderByArea(rings);
+    const polygons = groupNestedRings(orderedRings);
+
+    // If we got no polygons, we can infer that the values are either all above or all below the threshold.
+    // If everything is below, we shold add the entire bounding box as a polygon.
+    // see https://github.com/Turfjs/turf/issues/1797
+    if (polygons.length === 0 && matrix[0][0] < upperBand) {
+      const dx = matrix[0].length;
+      const dy = matrix.length;
+      polygons.push([
+        [
+          [0, 0],
+          [dx - 1, 0],
+          [dx - 1, dy - 1],
+          [0, dy - 1],
+          [0, 0],
+        ],
+      ]);
+    }
 
     contours.push({
-      groupedRings: groupedRings as Position[][][],
+      groupedRings: polygons,
       [property]: lowerBand + "-" + upperBand,
     });
+
+    prevSegments = segments;
   }
+
   return contours;
+}
+
+/**
+ * Run marching squares across the matrix and calculate the implied counterclockwise ordered line segments from each cell.
+ * @see https://en.wikipedia.org/wiki/Marching_squares for an visualization of the different cases
+ * @private
+ */
+function getSegments(
+  matrix: ReadonlyArray<ReadonlyArray<number>>,
+  threshold: number
+): [Position, Position][] {
+  const segments: [Position, Position][] = [];
+
+  const dx = matrix[0].length;
+  const dy = matrix.length;
+
+  for (let y = 0; y < dy - 1; y++) {
+    for (let x = 0; x < dx - 1; x++) {
+      const tr = matrix[y + 1][x + 1];
+      const br = matrix[y][x + 1];
+      const bl = matrix[y][x];
+      const tl = matrix[y + 1][x];
+
+      let grid =
+        (tl >= threshold ? 8 : 0) |
+        (tr >= threshold ? 4 : 0) |
+        (br >= threshold ? 2 : 0) |
+        (bl >= threshold ? 1 : 0);
+
+      switch (grid) {
+        case 0:
+          continue;
+        case 1:
+          segments.push([
+            [x + frac(bl, br), y],
+            [x, y + frac(bl, tl)],
+          ]);
+          break;
+        case 2:
+          segments.push([
+            [x + 1, y + frac(br, tr)],
+            [x + frac(bl, br), y],
+          ]);
+          break;
+        case 3:
+          segments.push([
+            [x + 1, y + frac(br, tr)],
+            [x, y + frac(bl, tl)],
+          ]);
+          break;
+        case 4:
+          segments.push([
+            [x + frac(tl, tr), y + 1],
+            [x + 1, y + frac(br, tr)],
+          ]);
+          break;
+        case 5: {
+          // use the average of the 4 corners to differentiate the saddle case and correctly honor the counter-clockwise winding
+          const avg = (tl + tr + br + bl) / 4;
+          const above = avg >= threshold;
+
+          if (above) {
+            segments.push(
+              [
+                [x + frac(tl, tr), y + 1],
+                [x, y + frac(bl, tl)],
+              ],
+              [
+                [x + frac(bl, br), y],
+                [x + 1, y + frac(br, tr)],
+              ]
+            );
+          } else {
+            segments.push(
+              [
+                [x + frac(tl, tr), y + 1],
+                [x + 1, y + frac(br, tr)],
+              ],
+              [
+                [x + frac(bl, br), y],
+                [x, y + frac(bl, tl)],
+              ]
+            );
+          }
+          break;
+        }
+        case 6:
+          segments.push([
+            [x + frac(tl, tr), y + 1],
+            [x + frac(bl, br), y],
+          ]);
+          break;
+        case 7:
+          segments.push([
+            [x + frac(tl, tr), y + 1],
+            [x, y + frac(bl, tl)],
+          ]);
+          break;
+        case 8:
+          segments.push([
+            [x, y + frac(bl, tl)],
+            [x + frac(tl, tr), y + 1],
+          ]);
+          break;
+        case 9:
+          segments.push([
+            [x + frac(bl, br), y],
+            [x + frac(tl, tr), y + 1],
+          ]);
+          break;
+        case 10: {
+          const avg = (tl + tr + br + bl) / 4;
+          const above = avg >= threshold;
+
+          if (above) {
+            segments.push(
+              [
+                [x, y + frac(bl, tl)],
+                [x + frac(bl, br), y],
+              ],
+              [
+                [x + 1, y + frac(br, tr)],
+                [x + frac(tl, tr), y + 1],
+              ]
+            );
+          } else {
+            segments.push(
+              [
+                [x, y + frac(bl, tl)],
+                [x + frac(tl, tr), y + 1],
+              ],
+              [
+                [x + 1, y + frac(br, tr)],
+                [x + frac(bl, br), y],
+              ]
+            );
+          }
+          break;
+        }
+        case 11:
+          segments.push([
+            [x + 1, y + frac(br, tr)],
+            [x + frac(tl, tr), y + 1],
+          ]);
+          break;
+        case 12:
+          segments.push([
+            [x, y + frac(bl, tl)],
+            [x + 1, y + frac(br, tr)],
+          ]);
+          break;
+        case 13:
+          segments.push([
+            [x + frac(bl, br), y],
+            [x + 1, y + frac(br, tr)],
+          ]);
+          break;
+        case 14:
+          segments.push([
+            [x, y + frac(bl, tl)],
+            [x + frac(bl, br), y],
+          ]);
+          break;
+        case 15:
+          // all above
+          continue;
+      }
+    }
+  }
+
+  return segments;
+
+  // get the linear interpolation fraction of how far z is between z0 and z1
+  // See https://github.com/fschutt/marching-squares/blob/master/src/lib.rs
+  function frac(z0: number, z1: number): number {
+    if (z0 === z1) {
+      return 0.5;
+    }
+
+    let t = (threshold - z0) / (z1 - z0);
+    return t > 1 ? 1 : t < 0 ? 0 : t;
+  }
+}
+
+/**
+ * Create a list of closed rings from the combined segments from breaks[i] and breaks[i-1].
+ * @private
+ */
+function assembleRings(
+  segments: Position[][],
+  matrix: number[][]
+): Position[][] {
+  const dy = matrix.length;
+  const dx = matrix[0].length;
+
+  const contours: Position[][] = [];
+  const result: Position[][] = [];
+
+  // Assemble contiguous line segments into contours. These are at least LineStrings,
+  // but for features that do not touch the edge of the matrix, they will actually wind up
+  // being an entirely closed LinearRing.
+  while (segments.length > 0) {
+    const contour: Position[] = [...segments.shift()!];
+    contours.push(contour);
+
+    let found: boolean;
+    do {
+      found = false;
+      for (let i = 0; i < segments.length; i++) {
+        const segment = segments[i];
+        // add the segment's end point to the end of the contour
+        if (
+          segment[0][0] === contour[contour.length - 1][0] &&
+          segment[0][1] === contour[contour.length - 1][1]
+        ) {
+          found = true;
+          contour.push(segment[1]);
+          segments.splice(i, 1);
+          break;
+        }
+        // add the segment's start point to the start of the contour
+        if (
+          segment[1][0] === contour[0][0] &&
+          segment[1][1] === contour[0][1]
+        ) {
+          found = true;
+          contour.unshift(segment[0]);
+          segments.splice(i, 1);
+          break;
+        }
+
+        // note that because the segments are all guaranteed to be counterclockwise,
+        // we do not join segment start to end of the contour or segment end to the start of contour
+      }
+
+      // if we reach here with found === false, that means that no remaining segments can be
+      // added to our contour. We begin again creating the next indepdenent contour.
+    } while (found);
+  }
+
+  // Now we loop again, taking the contours and ensuring that all of them are closed rings.
+  // Using segments from two different breaks[], and enforcing closed polygons are the
+  // two the major difference between the implementation of @turf/isolines and @turf/isobands.
+  while (contours.length > 0) {
+    const contour = contours[0];
+
+    // if a contour is closed, store it in the results and move to the next contour
+    if (
+      contour[0][0] === contour[contour.length - 1][0] &&
+      contour[0][1] === contour[contour.length - 1][1]
+    ) {
+      result.push(contour);
+      contours.shift();
+      continue;
+    }
+
+    // A contour that is not already closed is guaranteed to touch the bounding box of the matrix.
+    // We know that the polygon is ordered counter-clockwise, so we just need to follow
+    // the bounding box in a counterclockwise direction, looking for a contour to append.
+    // We may need to insert new positions along the corners, but we will eventually close the ring.
+
+    const end = contour[contour.length - 1];
+
+    let match: number;
+    let corner: Position;
+    if (end[0] === 0 && end[1] !== 0) {
+      // left side
+      match = getAdjacentContour(
+        contours,
+        (contour) => contour[0][0] === 0 && contour[0][1] < end[1], // left side, below end
+        (a, b) => b[0][1] - a[0][1] // prefer positions to the top
+      );
+      corner = [0, 0]; // bottom left corner
+    } else if (end[1] === 0 && end[0] !== dx - 1) {
+      // bottom side
+      match = getAdjacentContour(
+        contours,
+        (contour) => contour[0][1] === 0 && contour[0][0] > end[0], // bottom side, right of end
+        (a, b) => a[0][0] - b[0][0] // prefer positions to the left
+      );
+      corner = [dx - 1, 0]; // bottom right corner
+    } else if (end[0] === dx - 1 && end[1] !== dy - 1) {
+      // right side
+      match = getAdjacentContour(
+        contours,
+        (contour) => contour[0][0] === dx - 1 && contour[0][1] > end[1], // right side, above end
+        (a, b) => a[0][1] - b[0][1] // prefer positions to the bottom
+      );
+      corner = [dx - 1, dy - 1]; // top right corner
+    } else if (end[1] === dy - 1 && end[0] !== 0) {
+      // top side
+      match = getAdjacentContour(
+        contours,
+        (contour) => contour[0][1] === dy - 1 && contour[0][0] < end[0], // top side, left of end
+        (a, b) => b[0][0] - a[0][0] // prefer positions to the right
+      );
+      corner = [0, dy - 1]; // top left corner
+    } else {
+      throw new Error("Contour not closed but is not along an edge");
+    }
+
+    if (match === -1) {
+      // we did not match a contour on this side, so we add a point in the corner to
+      // continue creating our linestring in counterclockwise order. The next
+      // run of the loop will continue trying to assemble the current contour on the next side.
+      contour.push(corner);
+    } else if (match === 0) {
+      // We looped back to a contour, and it was ourself. That means that we finished closing the ring.
+      // Add the contour to the result and remove it from the contours list to start working
+      // on the next contour.
+      contour.push([contour[0][0], contour[0][1]]);
+      result.push(contour);
+      contours.shift();
+    } else {
+      // We matched a contour, but it is not the one we're currently closing.
+      // That means that we get to add its points to our own, and remove that contour entirely.
+      // On the next loop, we'll continue trying to close the same contour, but this time from
+      // the final Position in contour will be the end of contours[match].
+      const matchedContour = contours[match];
+      contours.splice(match, 1);
+      for (const p of matchedContour) {
+        contour.push(p);
+      }
+    }
+  }
+
+  // If we get *just* a corner we close it immediately with itself, which results in
+  // a 2 point 'ring', which has zero area. We omit these before returning.
+  for (let i = 0; i < result.length; i++) {
+    if (result[i].length < 4) {
+      result.splice(i, 1);
+      i--;
+    }
+  }
+
+  return result;
 }
 
 /**
@@ -279,6 +657,32 @@ function allGrouped(
     }
   }
   return true;
+}
+
+/**
+ * Utility function to help close contours into rings
+ *
+ * @private
+ * @param contours The list of contours
+ * @param test Return true if a contour is a candidate for being joined
+ * @param sort Compare two candidates, returning a positive number will swap the best match from a to b
+ * @returns An index of the contour to join, or -1 if no contour was found
+ */
+function getAdjacentContour(
+  contours: Position[][],
+  test: (contour: Position[]) => boolean,
+  sort: (a: Position[], b: Position[]) => number
+): number {
+  let match = -1;
+  for (let j = 0; j < contours.length; j++) {
+    if (test(contours[j])) {
+      if (match === -1 || sort(contours[match], contours[j]) > 0) {
+        match = j;
+      }
+    }
+  }
+
+  return match;
 }
 
 export { isobands };