--- a/packages/turf-isobands/marchingsquares.d.ts
+++ b/packages/turf-isobands/marchingsquares.d.ts
@@ -1,1 +1,0 @@
-declare module "marchingsquares";