--- a/packages/turf-isobands/package.json
+++ b/packages/turf-isobands/package.json
@@ -82,7 +82,6 @@
     "@turf/invariant": "workspace:*",
     "@turf/meta": "workspace:*",
     "@types/geojson": "^7946.0.10",
-    "marchingsquares": "^1.3.3",
     "tslib": "^2.8.1"
   }
 }