Because `marchingsquares` has a somewhat conflicted licensing story, I went ahead and reimplemented it from scratch to clear up any ambiguity.

Performance is slower than isolines, but overall we're quite a bit faster than the original implementation (2-3x again).

The matrix2 issues are fixed in [another PR](https://github.com/Turfjs/turf/pull/2925).

Fixes #1084, #2509

| Before | After |
|--------|-------|
|bigMatrix x 487 ops/sec ±1.27% (83 runs sampled)|bigMatrix x 1,358 ops/sec ±0.77% (94 runs sampled)|
|<img width="1190" height="474" alt="image" src="https://github.com/user-attachments/assets/c67186a0-9ffb-4e8e-a1f8-cb0412d6c546" />|<img width="1187" height="475" alt="image" src="https://github.com/user-attachments/assets/c4e3979c-5194-495d-a71a-3d79f61c1d1a" />|
|matrix1 x 44,815 ops/sec ±0.26% (98 runs sampled)|matrix1 x 84,368 ops/sec ±0.98% (97 runs sampled)|
|<img width="1012" height="836" alt="image" src="https://github.com/user-attachments/assets/ed6cee9c-51c3-4ed4-93b3-b338a08a2ace" />|<img width="1012" height="838" alt="image" src="https://github.com/user-attachments/assets/a48ce905-d6cc-4a8c-836d-969b1151de86" />|
|matrix2 x 22,873 ops/sec ±0.24% (99 runs sampled)|matrix2 x 47,717 ops/sec ±0.46% (100 runs sampled)|
|<img width="640" height="839" alt="image" src="https://github.com/user-attachments/assets/3b8c6941-2b03-45aa-824a-ff0ea492f5e0" />|<img width="638" height="839" alt="image" src="https://github.com/user-attachments/assets/310bb991-d851-4542-b9ef-c14ef21597bb" />|
|pointGrid x 28,737 ops/sec ±0.27% (97 runs sampled)|pointGrid x 55,908 ops/sec ±0.26% (99 runs sampled)|
|<img width="1008" height="841" alt="image" src="https://github.com/user-attachments/assets/bd031070-6b60-47ad-932d-94d76b9d9c84" />|<img width="1007" height="839" alt="image" src="https://github.com/user-attachments/assets/2a12b354-8a67-4fbb-bb50-9e1886668302" />|

Pretty sure this fixed #2509 as well. Here's a picture of the output I got, but since the test file is quite large I omitted adding it to permanent test cases since it isn't really clear what caused the issue in the first place.

<img width="1221" height="591" alt="image" src="https://github.com/user-attachments/assets/15b62bee-dc19-41e8-971e-4aacc80bbd24" />